+++ src/euphorie/client/browser/conditions.py
@@ -2,6 +2,7 @@
 from Acquisition import aq_inner
 from euphorie.client import CONDITIONS_VERSION
 from euphorie.client.model import get_current_account
+from plone import api
 from Products.Five import BrowserView
 
 import logging
@@ -21,17 +22,16 @@ class TermsAndConditions(BrowserView):
         return getattr(self.account, "tc_approved", None) is not None
 
     def __call__(self):
-        self.came_from = self.request.form.get("came_from")
-        if isinstance(self.came_from, list):
-            # If came_from is both in the querystring and the form data
-            self.came_from = self.came_from[0]
+        webhelpers = api.content.get_view(
+            name="webhelpers", context=self.context, request=self.request
+        )
+        self.came_from = webhelpers.get_came_from(
+            default=aq_inner(self.context).absolute_url()
+        )
 
         self.account = get_current_account()
         if self.request.environ["REQUEST_METHOD"] == "POST":
             self.account.tc_approved = CONDITIONS_VERSION
 
-            if self.came_from:
-                self.request.response.redirect(self.came_from)
-            else:
-                self.request.response.redirect(aq_inner(self.context).absolute_url())
+            self.request.response.redirect(self.came_from)
         return self.index()